archive/golden_armada/agno-agents/requirements.txt
agno
google-genai
python-dotenv
mcp
